+++ grails-app/conf/spring/resources.groovy
@@ -1,7 +1,7 @@
 /*
  * resources.groovy
  *
- * Copyright (c) 2011-2018, Daniel Ellermann
+ * Copyright (c) 2011-2022, Daniel Ellermann
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -26,6 +26,7 @@ import org.amcworld.springcrm.google.GoogleContactSync
 import org.amcworld.springcrm.google.GoogleContactSyncTask
 import org.amcworld.springcrm.install.diffset.NoteMarkdownDiffSet
 import org.amcworld.springcrm.install.diffset.ProjectDocumentDiffSet
+import org.amcworld.springcrm.install.diffset.PurchaseInvoiceNumberDiffSet
 import org.amcworld.springcrm.ldap.LdapFactory
 import org.amcworld.springcrm.xml.InvoicingTransactionXMLFactory
 import org.amcworld.springcrm.xml.LogErrorListener
@@ -57,6 +58,7 @@ beans = {
         markdownService = ref('markdownService')
     }
     startupDiffSet4(ProjectDocumentDiffSet)
+    startupDiffSet5(PurchaseInvoiceNumberDiffSet)
 
     /* XML and XSLT */
     fopFactory(FopFactory) { DefaultBeanConfiguration b ->