+++ grails-app/init/org/amcworld/springcrm/BootStrap.groovy
@@ -1,7 +1,7 @@
 /*
  * BootStrap.groovy
  *
- * Copyright (c) 2011-2016, Daniel Ellermann
+ * Copyright (c) 2011-2022, Daniel Ellermann
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -40,7 +40,7 @@ class BootStrap {
 
     //-- Constants ------------------------------
 
-    public static final int CURRENT_DB_VERSION = 4i
+    public static final int CURRENT_DB_VERSION = 6i
 
 
     //-- Fields ---------------------------------